requirements.txt
discord-webhook
requests
python-dotenv